done san pham

+++ src/main/java/com/example/be_datn/service/impl/SanPhamService.java
@@ -2,7 +2,6 @@ package com.example.be_datn.service.impl;
 
 import com.example.be_datn.dto.Request.SanPhamRequest;
 import com.example.be_datn.dto.Response.SanPhamResponse;
-import com.example.be_datn.entity.ChatLieuVai;
 import com.example.be_datn.entity.SanPham;
 import com.example.be_datn.exception.AppException;
 import com.example.be_datn.exception.ErrorCode;
@@ -31,8 +30,21 @@ public class SanPhamService implements ISanPhamService {
     }
 
     @Override
+    public SanPhamResponse updateStatus(Long idSanPham, int status) {
+        SanPham sanPham = sanPhamRepository.findById(idSanPham).orElseThrow(() -> new AppException(ErrorCode.SANPHAM_NOT_FOUND));
+        sanPham.setTrangThai(status);
+        return SanPhamResponse.fromSanPham(sanPhamRepository.save(sanPham));
+    }
+
+    @Override
+    public Page<SanPhamResponse> getAllWithFilter(Long idDanhMuc, Long idThuongHieu, Long idChatLieuVai, Long idChatLieuDe, String tenSanPham, Pageable pageable) {
+        return sanPhamRepository.getAllByFilter(idDanhMuc, idThuongHieu, idChatLieuVai, idChatLieuDe, "%" + tenSanPham + "%", pageable)
+                .map(SanPhamResponse::fromSanPham);
+    }
+
+    @Override
     public SanPhamResponse getById(Long id) {
-        return sanPhamRepository.findById(id).map(SanPhamResponse::fromSanPham).orElseThrow(() -> new ArithmeticException(ErrorCode.SANPHAM_NOT_FOUND.getMessage()));
+        return sanPhamRepository.findById(id).map(SanPhamResponse::fromSanPham).orElseThrow(() -> new AppException(ErrorCode.SANPHAM_NOT_FOUND));
     }
 
     @Override
@@ -40,7 +52,6 @@ public class SanPhamService implements ISanPhamService {
         if (sanPhamRepository.existsByTenSanPham(sanPhamRequest.getTenSanPham())) {
             throw new AppException(ErrorCode.TEN_SANPHAM_EXIST);
         }
-        // chưa test trường hợp chất liệu vải không có trong database vì chưa có code
         validateForeignKeys(sanPhamRequest);
 
         SanPham sanPham = buildSanPham(sanPhamRequest);
@@ -59,43 +70,54 @@ public class SanPhamService implements ISanPhamService {
 
     @Override
     public String delete(Long id) {
-        var sanPham = sanPhamRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.SANPHAM_NOT_FOUND));
-        sanPhamRepository.deleteById(id);
-        return "Deleted successfully";
+        if (sanPhamRepository.existsById(id)) {
+            sanPhamRepository.deleteById(id);
+            return "Deleted successfully";
+        } else {
+            throw new AppException(ErrorCode.SANPHAM_NOT_FOUND);
+        }
+
     }
 
 
     private void validateForeignKeys(SanPhamRequest request) {
-        chatLieuDeRepository.findById(request.getIdChatLieuDe())
-                .orElseThrow(() -> new AppException(ErrorCode.CHATLIEUDE_NOT_FOUND));
-
-        danhMucRepository.findById(request.getIdDanhMuc())
-                .orElseThrow(() -> new AppException(ErrorCode.DANHMUC_NOT_FOUND));
-
-        thuongHieuRepository.findById(request.getIdThuongHieu())
-                .orElseThrow(() -> new AppException(ErrorCode.THUONGHIEU_NOT_FOUND));
-        chatLieuVaiRepository.findById(request.getIdChatLieuVai()).orElseThrow(() -> new AppException(ErrorCode.CHATLIEUVAI_NOT_FOUND));
+        if (!chatLieuDeRepository.existsById(request.getIdChatLieuDe())) {
+            throw new AppException(ErrorCode.CHATLIEUDE_NOT_FOUND);
+        }
+        if (!danhMucRepository.existsById(request.getIdDanhMuc())) {
+            throw new AppException(ErrorCode.DANHMUC_NOT_FOUND);
+        }
+        if (!thuongHieuRepository.existsById(request.getIdThuongHieu())) {
+            throw new AppException(ErrorCode.THUONGHIEU_NOT_FOUND);
+        }
+        if (!chatLieuVaiRepository.existsById(request.getIdChatLieuVai())) {
+            throw new AppException(ErrorCode.CHATLIEUVAI_NOT_FOUND);
+        }
     }
 
     private void updateEntityFromRequest(SanPham sanPham, SanPhamRequest request) {
         sanPham.setTenSanPham(request.getTenSanPham())
                 .setMoTa(request.getMoTa())
                 .setTrangThai(request.getTrangThai())
-                .setDanhMuc(danhMucRepository.findById(request.getIdDanhMuc()).get())
-                .setChatLieuDe(chatLieuDeRepository.findById(request.getIdChatLieuDe()).get())
-                .setChatLieuVai(chatLieuVaiRepository.findById(request.getIdChatLieuVai()).get())
-                .setThuongHieu(thuongHieuRepository.findById(request.getIdThuongHieu()).get());
+                .setDanhMuc(danhMucRepository.findById(request.getIdDanhMuc())
+                        .orElseThrow(() -> new AppException(ErrorCode.DANHMUC_NOT_FOUND)))
+                .setChatLieuDe(chatLieuDeRepository.findById(request.getIdChatLieuDe())
+                        .orElseThrow(() -> new AppException(ErrorCode.CHATLIEUDE_NOT_FOUND)))
+                .setChatLieuVai(chatLieuVaiRepository.findById(request.getIdChatLieuVai())
+                        .orElseThrow(() -> new AppException(ErrorCode.CHATLIEUVAI_NOT_FOUND)))
+                .setThuongHieu(thuongHieuRepository.findById(request.getIdThuongHieu())
+                        .orElseThrow(() -> new AppException(ErrorCode.THUONGHIEU_NOT_FOUND)));
     }
 
     private SanPham buildSanPham(SanPhamRequest request) {
         return SanPham.builder()
                 .tenSanPham(request.getTenSanPham())
                 .moTa(request.getMoTa())
-                .ChatLieuVai(chatLieuVaiRepository.findById(request.getIdChatLieuVai()).get())
-                .ChatLieuDe(chatLieuDeRepository.findById(request.getIdChatLieuDe()).get())
-                .ThuongHieu(thuongHieuRepository.findById(request.getIdThuongHieu()).get())
+                .chatLieuVai(chatLieuVaiRepository.findById(request.getIdChatLieuVai()).orElseThrow(() -> new AppException(ErrorCode.CHATLIEUVAI_NOT_FOUND)))
+                .chatLieuDe(chatLieuDeRepository.findById(request.getIdChatLieuDe()).orElseThrow(() -> new AppException(ErrorCode.CHATLIEUDE_NOT_FOUND)))
+                .thuongHieu(thuongHieuRepository.findById(request.getIdThuongHieu()).orElseThrow(() -> new AppException(ErrorCode.THUONGHIEU_NOT_FOUND)))
+                .danhMuc(danhMucRepository.findById(request.getIdDanhMuc()).orElseThrow(() -> new AppException(ErrorCode.DANHMUC_NOT_FOUND)))
                 .trangThai(request.getTrangThai())
-                .DanhMuc(danhMucRepository.findById(request.getIdDanhMuc()).get())
                 .build();
     }
 }
